requirements.txt
cassandra-driver
locust
google-cloud-bigtable
google-cloud-core

